feat(aspectratio): convert to typescript
